migrate certificate related method from trainingService to certficateService
